% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_apexcharts.R
\name{apexchartOutput}
\alias{apexchartOutput}
\title{apexcharts output binding}
\usage{
apexchartOutput(inputId, width = "100\%", height = "100\%")
}
\value{
charts using apexchart js library
}
\description{
apexcharts output binding
}
